🚚 Move: 삭제 모달 및 헤더 스타일 코드 분리
