+++ src/components/organisms/Header/Header.jsx
@@ -1,12 +1,12 @@
 import { useState, useEffect, useRef, useContext } from 'react'
 import { Link, useNavigate } from 'react-router-dom'
-import { styled } from 'styled-components'
 import Logo from '../../../assets/Logo.svg'
 import ToggleBtn from '../../atoms/Button/ToggleBtn'
 import ColorIcon from '../../atoms/ColorIcon/ColorIcon'
 import LicatFace from '../../../assets/icon-liacat.svg'
 import { ResumeContext } from '../../../context/ResumeContext'
 import { MainBtn } from '../../atoms/Button'
+import * as styles from './Header-style'
 
 export default function Header({ options }) {
   const { resumeData } = useContext(ResumeContext)
@@ -36,8 +36,8 @@ export default function Header({ options }) {
   }, [menuRef])
 
   return (
-    <HeaderCont isWhite={isWhite}>
-      <HeaderLayout isCenter={isCenter}>
+    <styles.HeaderCont isWhite={isWhite}>
+      <styles.HeaderLayout isCenter={isCenter}>
         {isCenter ? (
           <h1>
             <Link to="/MAKE-RE_ver2/">
@@ -61,109 +61,35 @@ export default function Header({ options }) {
                 />
               </Link>
             </h1>
-            <BtnCont>
+            <styles.BtnCont>
               {hasCreate && (
                 <MainBtn type="create" onClick={moveWrite}>
                   이력서 만들기
                 </MainBtn>
               )}
               {hasProfile && (
-                <ProfileBtn onClick={handleToggleMenu} isMenuOpen={isMenuOpen}>
+                <styles.ProfileBtn
+                  onClick={handleToggleMenu}
+                  isMenuOpen={isMenuOpen}
+                >
                   <img alt="마이프로필" src={profileImg || LicatFace} />
                   {isMenuOpen && (
-                    <MenuList ref={menuRef}>
-                      <li>
+                    <styles.MenuList ref={menuRef}>
+                      <styles.MenuItem>
                         <Link to="/MAKE-RE_ver2/myresume">마이페이지</Link>
-                      </li>
-                      <li>
+                      </styles.MenuItem>
+                      <styles.MenuItem>
                         <button>로그아웃</button>
-                      </li>
-                    </MenuList>
+                      </styles.MenuItem>
+                    </styles.MenuList>
                   )}
-                </ProfileBtn>
+                </styles.ProfileBtn>
               )}
               <ToggleBtn />
-            </BtnCont>
+            </styles.BtnCont>
           </>
         )}
-      </HeaderLayout>
-    </HeaderCont>
+      </styles.HeaderLayout>
+    </styles.HeaderCont>
   )
 }
-
-const HeaderCont = styled.header`
-  width: 100vw;
-  background-color: ${(props) =>
-    props.isWhite ? 'var(--background-color)' : 'var(--gray-lv1-color)'};
-  border-bottom: ${(props) =>
-    props.isWhite ? '1px solid var(--gray-lv2-color)' : 'none'};
-`
-
-const HeaderLayout = styled.div`
-  max-width: 1190px;
-  height: 70px;
-  margin: 0 auto;
-  display: flex;
-  justify-content: ${(props) => (props.isCenter ? 'center' : 'space-between')};
-  align-items: center;
-`
-
-const BtnCont = styled.div`
-  display: flex;
-  align-items: center;
-`
-
-const ProfileBtn = styled.button`
-  position: relative;
-  width: 42px;
-  height: 42px;
-  border-radius: 50%;
-  border: 1px solid var(--gray-lv2-color);
-  background-color: var(--background-color);
-  border: ${(props) =>
-    props.isMenuOpen
-      ? '2px solid var(--primary-color)'
-      : '1px solid var(--gray-lv2-color)'};
-  margin-left: 20px;
-
-  img {
-    width: 100%;
-    height: 100%;
-    border-radius: 50%;
-  }
-`
-
-const MenuList = styled.ul`
-  position: absolute;
-  top: calc(100% + 8px);
-  right: 50%;
-  transform: translateX(50%);
-  width: 150px;
-  background-color: var(--background-color);
-  padding: 9px 8px;
-  border-radius: 10px;
-  box-shadow: 0px 5px 15px rgba(71, 73, 77, 0.1);
-  border: 1px solid var(--gray-lv2-color);
-  z-index: 100000;
-
-  li {
-    border-radius: 8px;
-    transition: background-color 0.1s ease-in;
-
-    a,
-    button {
-      color: var(--surface-color);
-      font-size: 14px;
-      display: inline-block;
-      width: 100%;
-      text-align: left;
-      line-height: 20px;
-      padding: 5px 10px;
-      border-radius: 8px;
-    }
-
-    &:hover {
-      background-color: var(--gray-lv1-color);
-    }
-  }
-`
